Add basic Node.js "Hello World" example
Created a basic Node.js application.
Added index.js with a simple "Hello, World!" message.
Set up package.json for project management.
Included a .gitignore file to exclude node_modules and other unnecessary files.
This commit sets up the foundation for a Node.js project, demonstrating how to output a simple message to the console. It's a starting point for learning and development with Node.js.